+++ libs/core/utils/helpers/src/lib/observable.helper.ts
@@ -16,6 +16,10 @@ import {
 import { ArrayHelper } from './array.helper';
 import { ObjectHelper } from './object.helper';
 
+type UnNull<T extends Record<string, unknown>> = {
+  [TKey in keyof T]: Exclude<T[TKey], null>;
+};
+
 export class ObservableHelper {
   static filterNullish<T>(): UnaryFunction<
     Observable<Nullable<T> | undefined>,
@@ -35,6 +39,22 @@ export class ObservableHelper {
   > {
     return pipe(
       filter((x) => x !== undefined) as OperatorFunction<T | undefined, T>
+    );
+  }
+
+  static filterNullishProp<T extends Record<string, unknown>>(
+    props: (keyof T)[]
+  ): UnaryFunction<Observable<T>, Observable<UnNull<T>>> {
+    return pipe(
+      filter((x) => {
+        for (let i = 0; i < props.length; i++) {
+          const prop = props[i];
+          if (ObjectHelper.isNullOrUndefined(x[prop])) {
+            return false;
+          }
+        }
+        return true;
+      }) as OperatorFunction<T, UnNull<T>>
     );
   }