+++ libs/web/teaching-schedule/ui/change-request/src/lib/change-request.component.ts
@@ -6,27 +6,15 @@ import {
   ViewChild,
 } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
-import { Store } from '@ngrx/store';
-import { TuiDestroyService } from '@taiga-ui/cdk';
 import { Teacher } from '@teaching-scheduling-system/web/shared/data-access/models';
-import {
-  AppShellState,
-  selectNotNullTeacher,
-} from '@teaching-scheduling-system/web/shared/data-access/store';
 import { NavbarService } from '@teaching-scheduling-system/web/shell/ui/navbar';
-import {
-  teachingScheduleRequestChangeOptions,
-  teachingScheduleRequestFilter,
-  teachingScheduleRequestReset,
-  TeachingScheduleRequestState,
-} from '@teaching-scheduling-system/web/teaching-schedule/data-access';
-import { Observable, takeUntil, tap } from 'rxjs';
+import { RequestStore } from '@teaching-scheduling-system/web/teaching-schedule/data-access';
+import { Observable, tap } from 'rxjs';
 
 @Component({
   templateUrl: './change-request.component.html',
   styleUrls: ['./change-request.component.css'],
   changeDetection: ChangeDetectionStrategy.OnPush,
-  providers: [TuiDestroyService],
 })
 export class ChangeRequestComponent implements AfterViewInit {
   // VIEWCHILD
@@ -38,31 +26,22 @@ export class ChangeRequestComponent implements AfterViewInit {
   // CONSTRUCTOR
   constructor(
     private readonly navbarService: NavbarService,
-    private readonly store: Store<TeachingScheduleRequestState>,
-    private readonly appShellStore: Store<AppShellState>,
-    private readonly destroy$: TuiDestroyService,
+    private readonly store: RequestStore,
     route: ActivatedRoute
   ) {
     const personal = route.snapshot.data['personal'] as boolean;
-    store.dispatch(teachingScheduleRequestReset({ personal }));
+    store.reset(personal);
 
-    this.teacher$ = this.appShellStore.pipe(
-      selectNotNullTeacher,
-      takeUntil(this.destroy$)
-    );
+    this.teacher$ = store.teacher$;
 
     if (personal) {
       this.handleSelectTeacher();
     } else {
-      this.store.dispatch(
-        teachingScheduleRequestFilter({
-          query: {
-            status: [],
-            page: 1,
-            pagination: 20,
-          },
-        })
-      );
+      store.filter({
+        status: [],
+        page: 1,
+        pagination: 20,
+      });
     }
   }
 
@@ -74,16 +53,7 @@ export class ChangeRequestComponent implements AfterViewInit {
   // PRIVATE METHODS
   private handleSelectTeacher(): void {
     this.teacher$
-      .pipe(
-        tap((teacher) => {
-          this.store.dispatch(
-            teachingScheduleRequestChangeOptions({
-              options: { teacher },
-            })
-          );
-        }),
-        takeUntil(this.destroy$)
-      )
+      .pipe(tap((teacher) => this.store.changeOptions({ teacher })))
       .subscribe();
   }
 }